refactor(i18n): streamline translation handling by loading translations at build time instead of bundling JSON files
