+++ .changeset/twenty-worms-attend.md
@@ -1,0 +1,5 @@
+---
+"studiocms": patch
+---
+
+Refactor: Handle i18n translations at build instead of bundling json files
