+++ .vscode/settings.json
@@ -182,6 +182,7 @@
     "studiosdk",
     "sublist",
     "suneditor",
+    "tête",
     "tinyexec",
     "tinyglobby",
     "tspc",
