+++ packages/studiocms/src/virtual.d.ts
@@ -122,6 +122,7 @@ declare module 'virtual:studiocms/components/Editors' {
 
 declare module 'studiocms:i18n/virtual' {
 	export const availableTranslationFileKeys: string[];
+	export const availableTranslations: typeof import('./virtuals/i18n/v-files').availableTranslations;
 }
 
 declare module 'studiocms:i18n' {
